Remove default credential expiration
Removes the expires time from JWT credentials we issue. Except for when
using the user data flow, and the provided user data JWT includes an
expiration time.

In that case, we use the customer's expiration time for the credential.
That is, the credential we issue will expire at the same time as the
provided user data JWT.

Why?

The original expiration time was a day. But, having the expiration
didn't actually _do_ anything useful. Once a user is in a room, they can
stay there forever until the connection is lost. The only thing the
expiration achieved is introducing odd states where a user could fail to
join a new room after a day.